Multi-target to .NET Standard 2.1, .NET 5 and .NET 6 (#69)
Adds targets for .NET Standard 2.1, .NET 5 and .NET 6 in order to specify better dependency versions. Previously, any users of the library would be forced to use the 6.0.0 version of all Microsoft libraries which might break compatibility with others. By using the dependency version suited to the .NET version (i.e. 3.1.22 for .NET Standard 2.1, 5.0.0 for .NET 5 and so on) this will allow the library to be as compatible with the wider ecosystem as possible.

Closes #68